Move code coverage target test to acceptance suite
Run the synthetic publish project through AcceptanceTestBase so the regression uses the repository's temporary-directory, patched-dotnet, process timeout, and output-capture conventions.


Copilot-Session: 1031e116-e457-42a0-9c9c-828732a012dd
